fix: change some type errors and move the datas to its own component
