+++ types/users.ts
@@ -1,15 +1,13 @@
 export interface User {
-    id: string;
-    name: string;
-    email: string;
-    phoneNumber: string;
-    joinedDate?: string;
-    location: string;
-    platform: string;
-    avatar: string;
-    date: string;
-  }
-  
-  export type SortField = 'name' | 'email' | 'joinedDate' | 'location';
-  
-  
+  id: string;
+  name: string;
+  email: string;
+  phoneNumber: string;
+  joinedDate?: string;
+  location: string;
+  platform?: string;
+  avatar?: string;
+  date?: string;
+}
+
+export type SortField = "name" | "email" | "joinedDate" | "location";
